Remove stale, unreferenced es_MX.po
This translation is low-quality and not as well maintained
as es.po, but even if it was, it makes no sense for people
to work on two translations to the same language. (The
terminology is largely the same, and spelling is identical
in both dialects — not at all like the Portugal–Brazil
split).