Further filter out scroll gestures from page flip gestures
Only trigger a page flip if a swipe gesture is long enough in the
horizontal direction. This is in addition to the existing restriction
where a swipe gesture won't trigger a page flip if it's too long in the
vertical direction. The end result avoids spurious page flips during
scrolling